+++ lib/pages/detail/detail_content.dart
@@ -1,0 +1,227 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:pola_flutter/models/company.dart';
+import 'package:pola_flutter/models/search_result.dart';
+import 'package:pola_flutter/i18n/strings.g.dart';
+import 'logotypes.dart';
+import 'expandandable_text.dart';
+import 'polish_capital_graph.dart';
+
+class DetailContent extends StatelessWidget {
+  DetailContent(this.searchResult);
+
+  final SearchResult searchResult;
+
+  @override
+  Widget build(BuildContext context) {
+    if (searchResult.companies == null) {
+      return Padding(
+        padding: EdgeInsets.all(8.0),
+        child: Text(searchResult.altText ?? ""),
+      );
+    }
+
+    final company = searchResult.companies!.first;
+    final score = company.plScore ?? 0;
+    final double plCapital = (company.plCapital ?? 0).toDouble();
+    final Translations t = Translations.of(context);
+    
+    final hasLogo = company.logotypeUrl != null;
+    final hasDescription = company.description?.isNotEmpty ?? false;
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 17.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 17.0),
+            child: Align(
+              alignment: Alignment.centerLeft,
+              child: Row(
+                children: [
+                  SvgPicture.asset(
+                    'assets/info.svg',
+                    height: 24.0,
+                    width: 24.0,
+                  ),
+                  SizedBox(width: 8.0),
+                  
+                  Text(
+                      t.companyScreen.ourRating,
+                    style: TextStyle(
+                      fontSize: 16.0,
+                      fontWeight: FontWeight.w600,
+                      fontFamily: 'Lato',
+                    ),
+                  ),
+                  SizedBox(width: 8.0),
+                  Text(
+                     t.companyScreen. points(score: score),
+                    style: TextStyle(
+                      fontSize: 20.0,
+                      fontWeight: FontWeight.w700,
+                      fontFamily: 'Lato',
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 17.0),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(10.0),
+              child: LinearProgressIndicator(
+                value: score / 100.0,
+                backgroundColor: Color(0xFFF5DEDD),
+                valueColor: AlwaysStoppedAnimation<Color>(Color(0xFFE1203E)),
+                minHeight: 12.0,
+              ),
+            ),
+          ),
+          SizedBox(height: 17.0),
+          Divider(
+            thickness: 1.0,
+            color: Color(0xFFF0F0F0),
+            indent: 17.0,
+            endIndent: 17.0,
+          ),
+          Padding(
+            padding: const EdgeInsets.only(left: 17.0, top: 22.0),
+            child: Align(
+              alignment: Alignment.centerLeft,
+              child: Text(
+                  t.companyScreen.gradingCriteria,
+                style: TextStyle(
+                  fontSize: 16.0,
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'Lato',
+                ),
+              ),
+            ),
+          ),
+          SizedBox(height: 22.0),
+          Row(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+               PolishCapitalGraph(percentage: plCapital),
+              SizedBox(width: 35.0),
+              Expanded(
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    _DetailItem(  t.companyScreen.producedInPoland, (company.plWorkers ?? 0) != 0),
+                    SizedBox(height: 14.0),
+                    _DetailItem(  t.companyScreen.researchInPoland, (company.plRnD ?? 0) != 0),
+                    SizedBox(height: 14.0),
+                    _DetailItem(  t.companyScreen.registeredInPoland, (company.plRegistered ?? 0) != 0),
+                    SizedBox(height: 14.0),
+                    _DetailItem(  t.companyScreen.notConcernPart, (company.plNotGlobEnt ?? 0) != 0),
+                  ],
+                ),
+              ),
+            ],
+          ),
+          SizedBox(height: 22.0),
+          Divider(
+            thickness: 1.0,
+            color: Color(0xFFF0F0F0),
+            indent: 17.0,
+            endIndent: 17.0,
+          ),
+          if (hasDescription) ...[
+            Padding(
+              padding: EdgeInsets.symmetric(vertical: 22.0, horizontal: 17.0),
+              child: ExpandableText(company.description ?? ""),
+            ),
+            if (hasLogo) Divider(
+              thickness: 1.0,
+              color: Color(0xFFF0F0F0),
+              indent: 17.0,
+              endIndent: 17.0,
+            ),
+          ],
+          if (hasLogo) Logotypes(logotypes: searchResult.logotypes(), searchResult: searchResult),
+          SizedBox(height: 26.0),
+          if (hasLogo) Divider(
+            thickness: 1.0,
+            color: Color(0xFFF0F0F0),
+            indent: 17.0,
+            endIndent: 17.0,
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class _DetailItem extends StatelessWidget {
+  _DetailItem(this.text, this.state);
+
+  final String text;
+  final bool state;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        Padding(
+          padding: EdgeInsets.only(right: 3.0),
+          child: SvgPicture.asset(
+            state ? 'assets/task_alt.svg' : 'assets/radio_button_unchecked.svg',
+          ),
+        ),
+        Expanded(
+          child: Text(
+            text,
+            style: TextStyle(
+              fontSize: 11,
+              fontWeight: FontWeight.w400,
+              fontFamily: 'Lato',
+            ),
+            softWrap: true,
+            overflow: TextOverflow.visible,
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+
+extension on SearchResult {
+  List<Logotype> logotypes() {
+    var brandLogotypes = allCompanyBrands?.map((brand) {
+          final brandLogotype = brand.logotypeUrl;
+          if (brandLogotype != null) {
+            return Logotype(brandLogotype, null);
+          } else {
+            return null;
+          }
+        }).toList() ??
+        [];
+
+    final logotypeCompany = companies?.first.logotype();
+
+    brandLogotypes.insert(0, logotypeCompany);
+
+    return brandLogotypes
+        .where((logotype) => logotype != null)
+        .cast<Logotype>()
+        .toList();
+  }
+}
+
+extension on Company {
+  Logotype? logotype() {
+    final logotypeUrl = this.logotypeUrl;
+    if (logotypeUrl != null) {
+      return Logotype(logotypeUrl, officialUrl);
+    } else {
+      return null;
+    }
+  }
+}
